Moved User samples to CommonSpec
