+++ src/test/scala/ru/skelantros/easymacher/services/UserServicesSpec.scala
@@ -15,14 +15,6 @@ import ru.skelantros.easymacher.utils.Email
 import ru.skelantros.easymacher.CommonSpec
 
 class UserServicesSpec extends AnyFlatSpec with CommonSpec {
-  val usersSample = Seq(
-    User(1, "skelantros", Email("skelantros@easymacher.ru").get, "23052001", Role.Admin, true, "001", Some("Alex"), Some("Egorowski")),
-    User(2, "adefful", Email("ad3fful@easymacher.ru").get, "1234", Role.User, true, "002", Some("Alex"), None),
-    User(3, "g03th3", Email("g03th3@klassik.de").get, "5678", Role.User, true, "003", None, None),
-    User(4, "damned", Email("damned@mail.ru").get, "xd", Role.User, false, "004", None, None)
-  )
-
-
   val skelantrosJson = Json.obj(
     "id" := 1,
     "username" := "skelantros",
